requirements.txt
Django==2.0.4
django-chamber==0.4.0
filemagic==1.6
PyPrind==2.9.9
pytz==2018.3
six==1.11.0
Unidecode==1.0.22
